WIP: Persist app state in SQLite instead of yaml
