+++ naturtag/storage/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa: F401
 from naturtag.storage.client import ImageSession, iNatDbClient
-from naturtag.storage.settings import Settings, UserTaxa
+from naturtag.storage.settings import Settings
 from naturtag.storage.setup import setup
+from naturtag.storage.user_taxa import UserTaxa
